Updated to feathers v4 and reconfigured authentication
