+++ client/package.json
@@ -6,9 +6,9 @@
     "@babel/core": "7.4.3",
     "@babel/plugin-proposal-decorators": "^7.1.6",
     "@babel/preset-react": "^7.0.0",
-    "@feathersjs/authentication-client": "^1.0.8",
-    "@feathersjs/feathers": "^3.3.1",
-    "@feathersjs/socketio-client": "^1.1.5",
+    "@feathersjs/authentication-client": "^4.5.1",
+    "@feathersjs/feathers": "^4.5.1",
+    "@feathersjs/socketio-client": "^4.5.1",
     "@svgr/webpack": "4.1.0",
     "@typescript-eslint/eslint-plugin": "1.6.0",
     "@typescript-eslint/parser": "1.6.0",
