+++ package.json
@@ -32,22 +32,22 @@
     "mocha": "mocha test/ --recursive --exit"
   },
   "dependencies": {
-    "@feathersjs/authentication": "^2.1.13",
-    "@feathersjs/authentication-client": "^1.0.8",
-    "@feathersjs/authentication-jwt": "^2.0.6",
-    "@feathersjs/authentication-local": "^1.2.7",
-    "@feathersjs/cli": "^3.9.0",
-    "@feathersjs/configuration": "^2.0.4",
-    "@feathersjs/errors": "^3.3.4",
-    "@feathersjs/express": "^1.2.7",
-    "@feathersjs/feathers": "^3.2.3",
-    "@feathersjs/socketio": "^3.2.7",
+    "@feathersjs/authentication": "^4.4.1",
+    "@feathersjs/authentication-client": "^4.5.1",
+    "@feathersjs/authentication-local": "^4.4.1",
+    "@feathersjs/authentication-oauth": "^4.4.1",
+    "@feathersjs/cli": "^4.2.5",
+    "@feathersjs/configuration": "^4.4.1",
+    "@feathersjs/errors": "^4.4.1",
+    "@feathersjs/express": "^4.4.1",
+    "@feathersjs/feathers": "^4.4.1",
+    "@feathersjs/socketio": "^4.4.1",
     "axios": "^0.18.0",
     "compression": "^1.7.3",
     "cors": "^2.8.4",
     "dotenv": "^6.1.0",
     "feathers-hooks-common": "^4.19.4",
-    "feathers-sequelize": "^3.1.2",
+    "feathers-sequelize": "^6.1.0",
     "helmet": "^3.13.0",
     "json-bigint": "^0.3.0",
     "localStorage": "^1.0.4",
